Integrated splash screen in TerasologyEngine
